+++ src/resources/slider/slider.service.ts
@@ -69,9 +69,12 @@ export class SliderService {
     });
     if (slidesToDelete.length)
       for (let i = 0; i < slidesToDelete.length; i++) {
-        const imageToDelete = slidesToDelete[i].image;
-        if (imageToDelete?.publicId)
-          await this.imagesService.delete(imageToDelete.publicId);
+        const slide = slidesToDelete[i];
+        if (slide) {
+          const imageToDelete = slide.image;
+          if (imageToDelete?.publicId)
+            await this.imagesService.delete(imageToDelete.publicId);
+        }
       }
 
     const formattedSlides = slides?.map(async ({ image, ...slide }) => ({
@@ -105,9 +108,9 @@ export class SliderService {
     const slides = await this.prisma.slide.findMany({
       where: { sliderConfigurationId: id },
     });
-    for (let i = 0; slides.length < i; i++) {
+    for (let i = 0; i < slides.length; i++) {
       const slide = slides[i];
-      if (slide.imageId) await this.imagesService.delete(slide.imageId);
+      if (slide && slide.imageId) await this.imagesService.delete(slide.imageId);
     }
 
     return await this.prisma.sliderConfiguration.delete({
